Add tests for container routes
